Many fixes everywhere

+++ Rewards.BLL/MedalsLogic.cs
@@ -1,17 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Rewards.BLL.Interface;
+﻿using Rewards.BLL.Interface;
 using Rewards.DAL.Interface;
 using Rewards.Entities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Rewards.BLL
 {
     public class MedalsLogic : IMedalsLogic
     {
+        #region Fields
         private readonly IMedalDao _medalDao;
+        #endregion
 
         #region Constructor
         public MedalsLogic(IMedalDao medalDao)
@@ -21,24 +21,62 @@ namespace Rewards.BLL
         #endregion
 
         #region Methods
-        public void Add(string name, string material)
+        public int Add(string name, string material)
         {
-            _medalDao.Add(name, material);
-        }
-
-        public void Delete(int id)
-        {
-            _medalDao.Delete(id);
+            if (String.IsNullOrEmpty(name) || String.IsNullOrEmpty(material))
+            {
+                throw new ArgumentNullException("It is not possible to ADD a medal without a name or material.");
+            }
+            else
+            {
+                return _medalDao.Add(name, material);
+            }
         }
 
         public void Update(int id, string name, string material)
         {
-            _medalDao.Update(id, name, material);
+            if (_medalDao.GetById(id) == null)
+            {
+                throw new Exception($"Medal with id = {id} not available");
+            }
+            else
+            {
+                if (String.IsNullOrEmpty(name) || String.IsNullOrEmpty(material))
+                {
+                    throw new ArgumentNullException("It is not possible to UPDATE a medal without a name or material.");
+                }
+                else
+                {
+                    _medalDao.Update(id, name, material);
+                }
+            }
+        }
+
+        public void Delete(int id)
+        {
+            if (_medalDao.GetById(id) == null || OccursInReward(id))
+            {
+                throw new Exception("Someone is awarded this medal or a medal does not exist."); 
+            }
+            else
+            {
+                _medalDao.Delete(id);
+            }
+        }
+
+        public Medal GetById(int id)
+        {
+            return _medalDao.GetById(id);
+        }
+
+        public bool OccursInReward(int id)
+        {
+            return _medalDao.OccursInReward(id);
         }
 
         public IEnumerable<Medal> GetAll()
         {
-            return _medalDao.GetAll();
+            return _medalDao.GetAll().ToList();
         }
         #endregion
     }
